Avoid json encoding hack
As long as the scanee implements `sql.Scanner`, the type of `src` is not
checked. Therefore, we can actually avoid encoding stuff to bytes and
just use the go map/list directly as a `driver.Value`.
However, I did bring in an additional runtime dependency `mapstructure`
which helps decode `map[string]any` to the given struct.