arrow_test: demonstrate segfault
When I construct an arrow batch that has lists of structs, and then I
select just one of them, I get a segfault in duckdb. Is this not allowed
for some reason? If I do `select *` it all works out. I notice that the
underlying API is marked as deprecated so it's hard to tell what's
allowed.