fix: switch Decimal methods to value receivers
Switches the Decimal methods from pointer to
value receivers.

The ScanType of DECIMAL columns is Decimal,
not *Decimal. Users would expect that the methods
are available on the exact type returned from ScanType. For example, this fix allows Decimal
to be used where fmt.Stringer is expected, e.g.
in the fmt.(S/F)Print(f) functions.